refactor(stelae): move the dolos-free residue out of the profile
The first of the four in-workspace steps toward extracting the Stelae
bounded context into its own repository. Pure mechanics: code that already
belongs to the protocol or to a profile-generic driver crosses the crate
boundary behind re-export shims, so every existing import path still
resolves and no CLI or test file changes.

Protocol residue into `crates/stelae`:

- `history_for` becomes `stelae::inscription::history_for`, beside the
  invariant it is the constructive half of, with its five tests.
- The fixed-arity canonical-CBOR decode helpers become `stelae::codec`.
  `Error` gains `MalformedRecord` and `HistoryBreak` to carry what they
  raise.

New crate `crates/stelae-driver`, with its own `Error` and the same
never-`dolos-*` boundary rule as `stelae`: `preflight`, `reporting`,
`Standing`, `scope_key` and the `digests` codec, each moved with its tests.

`dolos-snapshot` re-exports all of it at the old paths. The two `From`
implementations flatten the refusals that moved — `MalformedRecord`,
`OutOfOrder`, `NotEnoughSpace`, `MalformedInscription`, `HistoryBreak` —
back into the variants they had, so what a caller matches and what an
operator reads are unchanged; that is why there is no `Driver(..)` wrapper
variant.

Also drops the `registry` cargo feature from the CLI module's docs: no
manifest defines it, and the OCI surface rides `dolos-snapshot/oci`, which
the root manifest enables unconditionally.

`crates/snapshot/tests/goldens.rs` passes with zero re-pins, and
`cargo tree -e normal --all-features -p stelae -p stelae-driver` matches
nothing `^dolos(-|$)`.

Plan: plans/dolos-stelae-repo-split-residue-moves.md
