+++ crates/stelae-driver/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "stelae-driver"
+description = "Profile-generic driver machinery for the Stelae snapshot protocol"
+version.workspace = true
+edition.workspace = true
+
+[features]
+# Forwarded to `stelae/oci` so a driver built for a registry-capable binary and
+# the protocol crate under it agree on one feature, rather than each consumer
+# having to name both. Nothing here is `#[cfg]`-gated on it yet.
+oci = ["stelae/oci"]
+
+# The other half of the Stelae boundary from `stelae` itself: this crate holds
+# the lifecycle machinery a publisher or a restorer needs whatever it is
+# publishing, and like the protocol crate it must never depend on a `dolos-*`
+# package. `cargo tree -p stelae-driver -e normal --all-features` shows the
+# boundary holding. See adrs/004_stelae_snapshots.md, "Code layout".
+[dependencies]
+stelae = { path = "../stelae" }
+
+fs4.workspace = true
+minicbor.workspace = true
+serde_json.workspace = true
+thiserror.workspace = true
+tracing.workspace = true
+
+[dev-dependencies]
+tempfile = "3.20.0"
