+++ crates/stelae-driver/src/preflight.rs
